split syft export tests
